requirements.txt
streamlit
psycopg2-binary
pandas
numpy
scikit-learn
joblib
passlib[bcrypt]
python-jose[cryptography]
